requirements.txt
colorama==0.4.4
colorlog==4.7.2
datasets==2.15.0
evaluate==0.4.1
huggingface_hub==0.20.1
hyperopt==0.2.5
langchain==0.1.5
langchain_core==0.1.18
matplotlib==3.8.2
numpy==1.24.1
pandas==2.2.0
psutil==5.9.7
PyYAML==6.0.1
recbole==1.2.0
rouge_score==0.1.2
scikit_learn==1.3.2
timm==0.9.12
torch==2.1.2+cu118
torch_geometric==2.4.0
torchvision==0.16.2+cu118
tqdm==4.66.1
transformers==4.36.2
xgboost==2.0.3